Remove hashing StateFunc on platform credential and principal
When a APNS certificate is renewed and the private key remains
the same the new certificate is detected and an attempt is made
to update the SNS platform application. This fails because the
hashed value for private key (platform credential) is sent along
with the new certificate (platform principal) causing the AWS API
to reject the request.